chore: fixed pr comments
